refactor(routes): extract /team-settings into team_settings.js (god-file decomposition)
Route contract unchanged (284 routes), 0 /team-settings left inline, vitest 38 passed 268 passed . Lucy (GLM-5.2), driver-gated.
